+++ gastromanager-backend/src/main/java/com/kaiho/gastromanager/application/ingredient/dto/request/IngredientUpdateRequestDto.java
@@ -2,10 +2,12 @@ package com.kaiho.gastromanager.application.ingredient.dto.request;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record IngredientUpdateRequestDto(String name,
                                          String unit,
                                          String supplier,
-                                         double pricePerUnit,
+                                         BigDecimal pricePerUnit,
                                          int minimumStockQuantity) {
 }
